move applyUtopiaJSXComponentsChanges to project-file-utils
