+++ editor/src/core/workers/parser-printer/uid-fix.ts
@@ -11,42 +11,13 @@ import { isParseSuccess, ParsedTextFile, StaticElementPath } from '../../shared/
 import * as TP from '../../shared/template-path'
 import { setUtopiaIDOnJSXElement } from '../../shared/uid-utils'
 import { getUtopiaID, transformJSXComponentAtElementPath } from '../../model/element-template-utils'
-import { getComponentsFromTopLevelElements } from '../../model/project-file-utils'
+import {
+  applyUtopiaJSXComponentsChanges,
+  getComponentsFromTopLevelElements,
+} from '../../model/project-file-utils'
 import { mapArrayToDictionary } from '../../shared/array-utils'
 import { emptySet } from '../../shared/set-utils'
 import { fastForEach } from '../../shared/utils'
-
-// TODO move this function from editor-sate to a shared utils file, delete this copy-paste
-export function applyUtopiaJSXComponentsChanges(
-  topLevelElements: Array<TopLevelElement>,
-  newUtopiaComponents: Array<UtopiaJSXComponent>,
-): Array<TopLevelElement> {
-  // Run through the old top level elements, replacing the exported elements with those in the
-  // newly updated result with the same name.
-  // If it doesn't exist in the updated result, delete it.
-  // For any new items in the updated result, add them in.
-  const addedSoFar: Set<string> = emptySet()
-  let newTopLevelElements: Array<TopLevelElement> = []
-  fastForEach(topLevelElements, (oldTopLevelElement) => {
-    if (isUtopiaJSXComponent(oldTopLevelElement)) {
-      const updatedElement = newUtopiaComponents.find((e) => e.name === oldTopLevelElement.name)
-      if (updatedElement !== undefined) {
-        addedSoFar.add(updatedElement.name)
-        newTopLevelElements.push(updatedElement)
-      }
-    } else {
-      newTopLevelElements.push(oldTopLevelElement)
-    }
-  })
-
-  fastForEach(newUtopiaComponents, (updatedElement) => {
-    if (!addedSoFar.has(updatedElement.name)) {
-      newTopLevelElements.push(updatedElement)
-    }
-  })
-
-  return newTopLevelElements
-}
 
 export function fixParseSuccessUIDs(
   oldParsed: ParsedTextFile | null,
